# web_go_goa/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Calculator Service
    description: Service for multiplying numbers, a Goa teaser
    version: 0.0.1
servers:
    - url: http://localhost:8000
paths:
    /multiply/{a}/{b}:
        get:
            tags:
                - calc
            summary: multiply calc
            operationId: calc#multiply
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  schema:
                    type: integer
                    description: Left operand
                    example: 5401762099778430809
                    format: int64
                  example: 1918630006328122782
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  schema:
                    type: integer
                    description: Right operand
                    example: 4288748512599820841
                    format: int64
                  example: 4212629202012168060
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: integer
                                example: 8803302123552712831
                                format: int64
                            example: 1698882017578366363
    /openapi3.json:
        get:
            tags:
                - calc
            summary: Download ./gen/http/openapi3.json
            operationId: calc#/openapi3.json
            responses:
                "200":
                    description: File downloaded
    /openapi3.yaml:
        get:
            tags:
                - calc
            summary: Download ./gen/http/openapi3.yaml
            operationId: calc#/openapi3.yaml
            responses:
                "200":
                    description: File downloaded
components: {}
tags:
    - name: calc
      description: The calc service performs operations on numbers.